requirements.txt
fastapi
uvicorn
transformers
torch
faiss-cpu
pandas
numpy
pydantic
